LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/canvas/offscreen/compositing/2d.composite.globalAlpha.canvascopy.worker-expected.txt


FAIL 2d assert_approx_equals: Red channel of the pixel at (50, 25) expected 0 +/- 2 but got 125
